requirements.txt
requests
chromedriver-autoinstaller
selenium
pyyaml
bs4
beautifulsoup4